Work towards proper stream item creation and retrieval
Addition of (now defunct) StreamItemDateComparator for when the
retrieval was using a TreeMap (resorting the sorted data?), with tests
to rely on it.  Will remove in the next commit.

Added some fixture usage for the abstract server tests, as well as a
nested private impl for stream item data.  This cut down on clutter for
the test class.

Added the multiple item and date ordering tests, which are still not
passing. The insertion was changed to be correct and refactored into
several smaller methods, but the retrieval still needs work to make
tests pass.