Made all tests pass!
Tricky bug: needed to reverse the comparator so the priority queue
would work correctly, which was further complicated by the original
logic for adding an entity to a feed leaving out the edge to the first
entity in some cases.  Once this was fixed the algorithm needed to be
tweaked to appropriately handle entities with no streams.