Add Stream Entry Deletion
This is the beginning of a quest to start having tests clean up after
themselves in the db.
Add writer and reader tests for deletion.
Add deletion to the writer interface.
Add implementation of deletion in GraphServer.
Switch back to TinkerGraph for testing.
There seems to be a problem with feed retrieval affecting a test -
could feed ordering still have bugs?