Fixed Eclipse tests
Restlet can sometimes have issues in test environments - individual
tests can mistakenly share Restlet resources. In this case, the
CollabinateComponent used by certain tests was not created during that
test, it was created during other tests. Because the old instance was
missing an Admin engine, it was throwing a null reference. We could try
to fix Restlet or the Eclipse test engine, but we’ll just accept the
added dependency of always needing an Admin engine for now.